Fix misc virtual item issues (#5980)
