+++ Content.Client/Hands/Systems/HandsSystem.cs
@@ -193,9 +193,12 @@ namespace Content.Client.Hands
             hands.Gui.Update(new HandsGuiState(states, hands.ActiveHand));
         }
 
-        protected override void HandleContainerModified(EntityUid uid, SharedHandsComponent component, ContainerModifiedMessage args)
+        public override void UpdateHandVisuals(EntityUid uid, SharedHandsComponent? handComp = null, AppearanceComponent? appearance = null)
         {
-            base.HandleContainerModified(uid, component, args);
+            if (!Resolve(uid, ref handComp, ref appearance, false))
+                return;
+
+            base.UpdateHandVisuals(uid, handComp, appearance);
 
             if (uid == _playerManager.LocalPlayer?.ControlledEntity)
                 UpdateGui();
