+++ Content.Shared/Hands/Components/HandVirtualItemComponent.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using Robust.Shared.Containers;
 using Robust.Shared.GameObjects;
 using Robust.Shared.GameStates;
 using Robust.Shared.Players;
@@ -37,6 +38,10 @@ namespace Content.Shared.Hands.Components
                 return;
 
             _blockingEntity = pullState.BlockingEntity;
+
+            // update hands GUI with new entity.
+            if (Owner.TryGetContainer(out var containter))
+                EntitySystem.Get<SharedHandsSystem>().UpdateHandVisuals(containter.Owner);
         }
 
         [Serializable, NetSerializable]
